requirements.txt
# requirements.txt

# Core dependencies
requests==2.31.0
python-dotenv==1.0.0

# LLM Provider Libraries
openai==1.3.5
anthropic==0.7.4
google-generativeai==0.3.1
ollama==0.1.5

# Utility Libraries
tqdm==4.66.1
colorama==0.4.6  # For colored terminal output

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Type checking
mypy==1.7.1

# Logging
logging==0.4.9.6

# CLI enhancements
click==8.1.7  # For building command line interfaces

# Data handling
pandas==2.1.3  # For data manipulation and analysis
numpy==1.26.2  # For numerical operations

# Security
cryptography==41.0.5  # For handling API keys securely

# Development tools
black==23.11.0  # Code formatter
flake8==6.1.0  # Linter
isort==5.12.0  # Import sorter